some fixes

+++ app.py
@@ -4,7 +4,7 @@ from PIL import Image
 import numpy as np
 import custom.models
 from config import Cfg
-from production import get_predictions
+from production import get_predictions, window_image
 from zipfile import ZipFile
 import nibabel as nib
 import random
@@ -21,26 +21,30 @@ def download_model():
     gdown.cached_download(drive_link, quiet=False)
 
 
-def get_random_string():
+def get_folder_name():
     return ''.join(random.choice(string.ascii_lowercase) for _ in range(7)) + '/'
 
 
 def read_files(files):
-    folder_name = get_random_string()
+    folder_name = get_folder_name()
     path = 'images/' + folder_name
-    os.mkdir(path)
+    if not os.path.exists(path):
+        os.mkdir(path)
     imgs = list()
     for file in files:
         imgs.append([])
         if '.nii' in file.name:
             nii_path = path + file.name
             open(nii_path, 'wb').write(file.getvalue())
-            images = nib.load(nii_path)
+            try:
+                images = nib.load(nii_path)
+            except:
+                return None
             images = np.array(images.dataobj)
             images = np.moveaxis(images, -1, 0)
             os.remove(nii_path)
             
-            for idx, image in enumerate(images):
+            for image in images:
                 image = window_image(image, -600, 1500)
                 image += abs(np.min(image))
                 image = image / np.max(image)
@@ -55,15 +59,6 @@ def read_files(files):
 
             imgs[-1].append(path + file.name)
     return imgs, folder_name
-
-
-def window_image(image, window_center, window_width):
-    img_min = window_center - window_width // 2
-    img_max = window_center + window_width // 2
-    window_image = image.copy()
-    window_image[window_image < img_min] = img_min
-    window_image[window_image > img_max] = img_max
-    return window_image
 
 
 def main():
@@ -92,8 +87,10 @@ def main():
         unsafe_allow_html=True,
     )
     # download_model()
-    if not os.path.exists('segmentations/'):
-        os.mkdir('segmentations/')
+    for folder in ['segmentations/', 'images/']:
+        if not os.path.exists(folder):
+            os.mkdir(folder)
+    
 
     st.title('Сегментация поражения легких коронавирусной пневмонией')
 
@@ -107,52 +104,55 @@ def main():
         print(filenames)
         images, folder_name = read_files(filenames)
         
-        user_dir = "segmentations/" + folder_name
-        os.mkdir(user_dir)
-        
-        cfg = Cfg(multi_class)
-        zip_obj = ZipFile('segmentations.zip', 'w')
-        with st.expander("Информация о каждом фото"):
-            info = st.info('Делаем предсказания, пожалуйста, подождите')
-            for image_list in images:
-                for filename, pred in zip(image_list[:2], get_predictions(cfg, image_list[:2])):    
-                    info.empty()
-                    st.markdown(f'<h3>{filename.split("/")[-1]}</h3>', unsafe_allow_html=True)
-
-                    original = np.array(Image.open(filename))
-                    col1, col2 = st.columns(2)
-                    col1.header("Оригинал")
-                    col1.image(original, width=350)
-                    
-                    to_img = pred * 255
-                    image_path = user_dir + filename.split('/')[-1]
-                    cv2.imwrite(image_path, to_img)
-                    zip_obj.write(image_path)
-                    
-                    col2.header("Сегментация")
-                    col2.image(pred, width=350)
-
-                    st.markdown('<br />', unsafe_allow_html=True)
-
-            zip_obj.close()
+        if not images:
+            st.error('Неправильный формат или название файла')
+        else:
+            user_dir = "segmentations/" + folder_name
+            os.mkdir(user_dir)
             
-        with st.expander("Скачать сегментации"):
-            with open('segmentations.zip', 'rb') as file:
-                st.download_button(
-                    label="Архив сегментаций",
-                    data=file,
-                    file_name="segmentations.zip",
-            )
-        
-        for file in os.listdir(user_dir):
-            os.remove(user_dir + file)
+            cfg = Cfg(multi_class)
+            zip_obj = ZipFile(user_dir + 'segmentations.zip', 'w')
+            with st.expander("Информация о каждом фото"):
+                info = st.info('Делаем предсказания, пожалуйста, подождите')
+                for image_list in images:
+                    for filename, pred in zip(image_list[:2], get_predictions(cfg, image_list[:2])):    
+                        info.empty()
+                        st.markdown(f'<h3>{filename.split("/")[-1]}</h3>', unsafe_allow_html=True)
+
+                        original = np.array(Image.open(filename))
+                        col1, col2 = st.columns(2)
+                        col1.header("Оригинал")
+                        col1.image(original, width=350)
+                        
+                        to_img = pred * 255
+                        image_path = user_dir + filename.split('/')[-1]
+                        cv2.imwrite(image_path, to_img)
+                        zip_obj.write(image_path)
+                        
+                        col2.header("Сегментация")
+                        col2.image(pred, width=350)
+
+                        st.markdown('<br />', unsafe_allow_html=True)
+
+                zip_obj.close()
+                
+            with st.expander("Скачать сегментации"):
+                with open('segmentations.zip', 'rb') as file:
+                    st.download_button(
+                        label="Архив сегментаций",
+                        data=file,
+                        file_name="segmentations.zip",
+                )
             
-        image_dir = 'images/' + folder_name
-        for file in os.listdir(image_dir):
-            os.remove(image_dir + file)
-            
-        os.rmdir(user_dir)
-        os.rmdir(image_dir)
+            for file in os.listdir(user_dir):
+                os.remove(user_dir + file)
+                
+            image_dir = 'images/' + folder_name
+            for file in os.listdir(image_dir):
+                os.remove(image_dir + file)
+                
+            os.rmdir(user_dir)
+            os.rmdir(image_dir)
